Update script.js
Swapped alt for ctrl to enable message download. Now renamed to ChatStatsLog. Made button change appearance when mode is switched. Changed ChatStatsLog to a txt file in the same format as normal chat logs, with an additional column for a symbol art indicator.

Enabled the importing of ChatStatLog files instead or in addition to normal logs.

Also implemented checking for duplicate messages and converting all times encountered into utc times.